Handle secrets.
Secrets are in their own little TOML file for development purposes, with
a short bit in the README about how to get set up. My overall idea is
for example data to be slurped down using curl to files so that the
build and validation process is offline, sandboxed, and repeatable. New
changes to ESI can be fetched manually by a developer, and there may be
some kind of obfuscator to sanitize the non-public parts of the payload.